requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.10
psycopg2-binary==2.9.6
redis==4.4.1
aioredis==2.0.1
pydantic==1.10.2  # Alterado para uma versão compatível com o FastAPI
python-dotenv==0.21.0
asyncpg==0.27.0
celery==5.2.7
pytest==7.2.2
pytest-asyncio==0.20.3
httpx==0.23.0
aiohttp==3.8.1
ulid-py==1.1.0
passlib==1.7.4
bcrypt==3.2.0
pyjwt==2.4.0


